Akbank/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
chromadb==0.4.22
streamlit==1.30.0
python-dotenv==1.0.0
bcrypt==4.1.2
google-generativeai==0.3.2